Add safety checks for view of Vector
Before numpy views of the Vector would persist even if the vector
reallocated itself. This could lead to segmentation faults if
they were accessed or written to.

Now if there is a view that has not been garbage collected, the
Vector is not able to be edited (insertion or deletion).

It might be possible to invalidate the view reference, but I
am not sure how to do that.