Fix serialization of samplesets with non-integer sample values
Sample sets allows the samples to be stored as arbitrary values
in order to avoid unnecessary copies. Unfortunately the
serialization method relied on them being integers or booleans.

The logic is now fixed to create a new boolean array in the case
that the samples are not already 0/1 and integer or boolean.