Generate adjvectorbqm.pyx and adjmapbqm.pyx from template
Because AdjVectorBQM and AdjMapBQM use the same set of overloaded
c++ functions to implement their logic, they have largely identical
source code. We would like to use fused types to handle this case,
but Cython, as of 0.29.13, does not support fused types as attributes
to cdef classes. To get around this, we generate the .pyx files from
a template and have two differnt .pxd files for the two bqm types.