Add a new clearer implementation of FileView
This implementation is inspired by the BufferedIOBase in that is
has a .readinto1 method. This means we always get the data from
a full neighborhood, then cut down that neighborhood if necessary.

The new version does a binary search every single time there is
a .readinto1 call, but in practice the new implementation seems
to come with only about a 50% slowdown relative to the former
implementation. Because the primary use for this feature is with
network uploads where the network latency dominates, this is
not really an issue.

In the future, one could overwrite the FileView.pos setter/getter
to track which "section" the pos is in, as well as increment
variables, positions, etc.