Backport dependence on dwave-preprocessing
This is to enable dwave-system (>1.7.0) to support both dimod 0.9 and
0.10.

Although now we have duplicated composites (available in both dimod
and dwave.preprocessing namespaces), this is necessary to resolve
unsatisfiable dependencies when dwave-system 1.7.x is installed with
dimod 0.9.x (due to dwave-preprocessing introduced as extras dependency
in dimod 0.10.0 and then removed in 0.10.5). This makes specifying dimod
as a dependency in dwave-system impossible, if we want to allow both 0.9
and 0.10.